Define moveToFront() and swapFrontward() functions
Formally defines and tests the moveToFront() and swapFrontward() functions, which take a 1D array and an index n such that 1 ≤ n ≤ length(array) and return the modifed array in which the nth element is either moved to the front of the array (i.e. to index = 1) or the nth element is swapped with the (n-1)st element. If n = 1, both functions simply return the original array.